refactor(src): 페이지에 'Page' suffix 추가
